unit tests for surface patches moved to separate test module
